Added npc spawning

+++ ConquerEmulator/ConquerBoxGame/src/conquerboxgame/structures/ChatTypes.java
@@ -26,32 +26,32 @@ package conquerboxgame.structures;
  */
 public class ChatTypes
 {
-    public final static int ACTION                = 0x7d2;
-    public final static int BROADCAST             = 0x7da;
-    public final static int CENTER                = 0x7db;
-    public final static int DIALOG                = 0x834;
-    public final static int FRIEND                = 0x7d9;
-    public final static int FRIENDBOARD           = 0x89a;
-    public final static int FRIENDSOFFLINEMESSAGE = 0x83e;
-    public final static int GHOST                 = 0x7dd;
-    public final static int GUILD                 = 0x7d4;
-    public final static int GUILDBOARD            = 0x89c;
-    public final static int GUILDBULLETIN         = 0x83f;
-    public final static int LOGININFORMATION      = 0x835;
-    public final static int MINIMAP               = 0x83c;
-    public final static int MINIMAP2              = 0x83d;
-    public final static int OTHERSBOARD           = 0x89d;
-    public final static int SERVICE               = 0x7de;
-    public final static int SPOUSE                = 0x7d6;
-    public final static int TALK                  = 0x7d0;
-    public final static int TEAM                  = 0x7d3;
-    public final static int TEAMBOARD             = 0x89b;
-    public final static int TOP                   = 0x7d5;
-    public final static int TRADEBOARD            = 0x899;
-    public final static int VENDORHAWK            = 0x838;
-    public final static int WEBSITE               = 0x839;
-    public final static int WHISPER               = 0x7d1;
-    public final static int YELL                  = 0x7d8;
+    public final static short ACTION                = 0x7d2;
+    public final static short BROADCAST             = 0x7da;
+    public final static short CENTER                = 0x7db;
+    public final static short DIALOG                = 0x834;
+    public final static short FRIEND                = 0x7d9;
+    public final static short FRIENDBOARD           = 0x89a;
+    public final static short FRIENDSOFFLINEMESSAGE = 0x83e;
+    public final static short GHOST                 = 0x7dd;
+    public final static short GUILD                 = 0x7d4;
+    public final static short GUILDBOARD            = 0x89c;
+    public final static short GUILDBULLETIN         = 0x83f;
+    public final static short LOGININFORMATION      = 0x835;
+    public final static short MINIMAP               = 0x83c;
+    public final static short MINIMAP2              = 0x83d;
+    public final static short OTHERSBOARD           = 0x89d;
+    public final static short SERVICE               = 0x7de;
+    public final static short SPOUSE                = 0x7d6;
+    public final static short TALK                  = 0x7d0;
+    public final static short TEAM                  = 0x7d3;
+    public final static short TEAMBOARD             = 0x89b;
+    public final static short TOP                   = 0x7d5;
+    public final static short TRADEBOARD            = 0x899;
+    public final static short VENDORHAWK            = 0x838;
+    public final static short WEBSITE               = 0x839;
+    public final static short WHISPER               = 0x7d1;
+    public final static short YELL                  = 0x7d8;
 }
